backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
alabaster==1.0.0
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-libsql==0.1.3
djangorestframework==3.17.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
docutils==0.22.4
frozenlist==1.8.0
idna==3.11
imagesize==2.0.0
Jinja2==3.1.6
libsql-client==0.3.1
MarkupSafe==3.0.3
multidict==6.7.1
mysqlclient==2.2.8
packaging==26.0
propcache==0.4.1
Pygments==2.20.0
PyJWT==1.7.1
python-dotenv==1.2.2
requests==2.33.1
roman-numerals==4.1.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
